refactor: generalize task details variables view component
