+++ tasklist/client/src/common/tasks/variables-editor/JSONEditorModal/index.tsx
