+++ tasklist/client/src/common/tasks/variables-editor/JSONEditorModal/styles.module.scss
