+++ tasklist/client/src/common/tasks/variables-editor/ResetForm.tsx
