+++ tasklist/client/src/common/tasks/variables-editor/VariableEditor/OnNewVariableAdded.tsx
