+++ tasklist/client/src/common/tasks/variables-editor/VariableEditor/styles.module.scss
