+++ tasklist/client/src/common/tasks/variables-editor/styles.module.scss
